Add files via upload
/*
The Aim is to create a platform which has the capability to interact with user who are in need of a car on rent. It also works as an intermediate between users where they agrees to exchange their cars. To acquire the facilities, every one has to login to the application, then a person looking for a car gets the access to the list of cars available with rent details and can book any of then he wishes to.
A person looking for customers is registered and details of his car are taken to make it available to be booked.
*/